Final commit challenge 15 display similar boats - see message
targetConfig(s) elements, each specifying settings for a particular target where the component can be used. In this example, the target is lightning_page
which means these settings apply when the component is added to a record page

datasource: A comma-separated list of values that the property can take. This is useful for creating a picklist of predefined options.
type: The data type of the property. Common types include String Boolean Integer etc
name: he name of the property. This is how you reference the property in your JavaScript